backend/requirements.txt
asgiref==3.8.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
boto3==1.34.84
botocore==1.34.162
cbor2==5.9.0
certifi==2026.4.22
cffi==2.0.0
channels==4.0.0
charset-normalizer==3.4.7
constantly==23.10.4
cryptography==46.0.7
daphne==4.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jmespath==1.1.0
msgpack==1.1.2
packaging==24.0
pillow==10.3.0
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.8.0
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
requests==2.33.1
s3transfer==0.10.4
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.0
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
whitenoise==6.6.0
zope.interface==8.3
